extract RegionExecutor
